+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/command_attachment/command/argument/wrapper/InteractType.java
@@ -10,5 +10,7 @@ public enum InteractType {
     @SerializedName(value = "ANY_CLICK", alternate = "BOTH")
     ANY_CLICK,
 
+    SWAP_HAND,
+
     STEP_ON
 }